Support 64-bit integers (longs) in LuaState.

+++ src/main/java/li/cil/repack/com/naef/jnlua/DefaultJavaReflector.java
@@ -376,14 +376,14 @@ public class DefaultJavaReflector implements JavaReflector {
 							"attempt to read array with %s accessor",
 							luaState.typeName(2)));
 				}
-				int index = luaState.toInteger(2);
+				long index = luaState.toInteger(2);
 				int length = Array.getLength(object);
 				if (index < 1 || index > length) {
 					throw new LuaRuntimeException(String.format(
 							"attempt to read array of length %d at index %d",
 							length, index));
 				}
-				luaState.pushJavaObject(Array.get(object, index - 1));
+				luaState.pushJavaObject(Array.get(object, (int) (index - 1)));
 				return 1;
 			}
 
@@ -424,7 +424,7 @@ public class DefaultJavaReflector implements JavaReflector {
 							"attempt to write array with %s accessor",
 							luaState.typeName(2)));
 				}
-				int index = luaState.toInteger(2);
+				long index = luaState.toInteger(2);
 				int length = Array.getLength(object);
 				if (index < 1 || index > length) {
 					throw new LuaRuntimeException(String.format(
@@ -440,7 +440,7 @@ public class DefaultJavaReflector implements JavaReflector {
 									luaState.typeName(3)));
 				}
 				Object value = luaState.toJavaObject(3, componentType);
-				Array.set(object, index - 1, value);
+				Array.set(object, (int) (index - 1), value);
 				return 0;
 			}
 
@@ -676,11 +676,11 @@ public class DefaultJavaReflector implements JavaReflector {
 			public int invoke(LuaState luaState) {
 				List<?> list = luaState.checkJavaObject(1, List.class);
 				int size = list.size();
-				int index = luaState.checkInteger(2);
+				long index = luaState.checkInteger(2);
 				index++;
 				if (index >= 1 && index <= size) {
 					luaState.pushInteger(index);
-					luaState.pushJavaObject(list.get(index - 1));
+					luaState.pushJavaObject(list.get((int) (index - 1)));
 					return 2;
 				} else {
 					luaState.pushNil();
@@ -696,11 +696,11 @@ public class DefaultJavaReflector implements JavaReflector {
 			public int invoke(LuaState luaState) {
 				Object array = luaState.checkJavaObject(1, Object.class);
 				int length = java.lang.reflect.Array.getLength(array);
-				int index = luaState.checkInteger(2);
+				long index = luaState.checkInteger(2);
 				index++;
 				if (index >= 1 && index <= length) {
 					luaState.pushInteger(index);
-					luaState.pushJavaObject(Array.get(array, index - 1));
+					luaState.pushJavaObject(Array.get(array, (int) (index - 1)));
 					return 2;
 				} else {
 					luaState.pushNil();
